# Pickle object-ified repodata for performance