Builds on #4473. By moving the IndexRecord wrapping into the fetch_repodata tasks, we can cache those objects. Significant improvements in speed result for subsequent loads of the same channels.

For `defaults` and `conda-forge`: first call to `get_index`, no pickled data:
```
CPU times: user 7.5 s, sys: 209 ms, total: 7.71 s
Wall time: 7.79 s
```
Second call:
```
CPU times: user 834 ms, sys: 86.6 ms, total: 921 ms
Wall time: 1.11 s
```
The cost is, in my view, reasonable. Here is the equivalent call prior to this PR:
```
CPU times: user 6.6 s, sys: 127 ms, total: 6.73 s
Wall time: 6.94 s
```