@mcg1969 2017-01-29
Oh, and here's 4.2, reading from cache.
```
CPU times: user 2.81 s, sys: 177 ms, total: 2.98 s
Wall time: 2.99 s
```