@mcg1969 2017-01-30
@kalefranz this is failing on the conda-build test. I'm not sure why, and of course it's essential that this pass. Nevertheless I really think this is a promising option, even if we make it optional.