@mcg1969 2017-01-30
Found the problem.