@kalefranz 2017-01-30
This is awesome progress.  I think the only thing I'd really want to add is a configuration parameter that turns the pickling off by default.  Probably `context.use_pickled_repodata`.  Just add a line in `conda/base/context.py` to the `Context` object

    use_pickled_repodata = PrimitiveParameter(False)

Then Ray and whoever else can enable it by setting condarc or a `CONDA_USE_PICKLED_REPODATA=true` environment variable.  We can also set that environment variable for the conda-build test run.