@kalefranz 2017-01-30
I'm 👍 👍 on this PR.  And I'm comfortable rolling it out without the extra configuration, if we can be just a little more conservative on the logic for when we don't use the pickled data.  I think we can get it merged today.