@kalefranz 2017-01-30
I added this `Priority` object that we can use as a memory reference.  Let's see what the CI does with the total time.  I just wanted to see what it would look like.  We can take it back out if you think it's too risky.

Note though that we're still just fine for `supplement_index_with_prefix()` and supplement_index_with_cache()` working correctly.  When the priority attribute is reassigned, it drops the reference `Priority` object reference.