@mcg1969 2017-01-30
I don't know what the tests are showing, but if there's not a significant performance benefit to your `Priority` trick, I think we should take it out and stick with a linear scan. It's definitely simpler and more clear.