@kalefranz 2017-01-31
Tests show stick with linear scan, but skip the rewrite-to-disk on change.  I say that's what we go with.